refactor(turn-runner): move the await-edge store tests into their own file
store.rs had grown to 1,046 lines with the `#[cfg(test)]` module starting at
472 — over half the file was tests, and the file crossed the repo's 1,000-line
ceiling for touched source. Split the test module into `store/tests.rs`,
matching the idiom already used three times in this crate
(`structured_finalization/tests.rs`, `loop_driver_host/compaction_tests.rs`,
`loop_driver_host/run_lease_fence_tests.rs`).

Pure move: the production half of `store.rs` is byte-identical to before, and
the test body differs only by the rustfmt reflow that follows dedenting it one
level. No behavior delta. `store.rs` is now 473 lines.
